Update go/kafka integration
Working on making Go code correctly interface with Kafka.  The issue is that we're trying to access Kafka before Kafka becomes available.  Trying to make the Go api container depend on the Kafka container does not seem to work.  Instead, I'm waiting 10 seconds before accessing Kafka, which seems to work but is not pretty in any shape or form.